+++ src/app/features/reservation/reservation-detail/reservation-detail.component.ts
@@ -7,7 +7,9 @@ import { CarService } from '../../../../generated/api/car.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { takeUntil } from 'rxjs/operators';
 import { ReservationService } from '../../../../generated/api/reservation.service';
-import { DatePipe, DecimalPipe } from '@angular/common';
+import { DecimalPipe } from '@angular/common';
+import { MatDialog } from '@angular/material/dialog';
+import { YesNoDialogComponent } from '../../../shared/yes-no-dialog/yes-no-dialog.component';
 
 @Component({
   selector: 'cr-reservation-detail',
@@ -37,7 +39,8 @@ export class ReservationDetailComponent implements OnInit, OnDestroy {
               private reservationService: ReservationService,
               private carService: CarService,
               private customerService: CustomerService,
-              private snackBar: MatSnackBar) {
+              private snackBar: MatSnackBar,
+              private dialog: MatDialog) {
   }
 
   ngOnInit() {
@@ -131,14 +134,24 @@ export class ReservationDetailComponent implements OnInit, OnDestroy {
   }
 
   public deleteReservation() {
-    this.reservationService.deleteReservationById(this.reservation.idReservation)
-      .pipe(takeUntil(this._onDestroy))
-      .subscribe(() => {
-        this.snackBar.open('Reservation löschen erfolgreich.', 'OK', {duration: 2000, panelClass: 'cr-snackbar-success'});
-        this.router.navigate(['/reservation']);
-      }, () => {
-        this.snackBar.open('Reservation löschen fehlgeschlagen.', 'OK', {duration: 2000, panelClass: 'cr-snackbar-error'});
-      });
+
+    const dialogRef = this.dialog.open(YesNoDialogComponent, {
+      width: '400px',
+      data: {title : 'Reservation löschen', text: 'Wollen Sie die Reservation wirklich löschen?'}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.reservationService.deleteReservationById(this.reservation.idReservation)
+          .pipe(takeUntil(this._onDestroy))
+          .subscribe(() => {
+            this.snackBar.open('Reservation löschen erfolgreich.', 'OK', {duration: 2000, panelClass: 'cr-snackbar-success'});
+            this.router.navigate(['/reservation']);
+          }, () => {
+            this.snackBar.open('Reservation löschen fehlgeschlagen.', 'OK', {duration: 2000, panelClass: 'cr-snackbar-error'});
+          });
+      }
+    });
   }
 
   ngOnDestroy() {
